Add an integration test for MaxFileCount
