+++ cloud/filestore/tests/loadtest/service-kikimr-test/max-file-count.txt
@@ -1,0 +1,25 @@
+Tests {
+    LoadTest {
+        Name: "max-file-count"
+        CreateFileStoreRequest {
+            FileSystemId: "max-file-count"
+            FolderId: "folder"
+            CloudId: "cloud"
+            BlocksCount: 10241024
+            BlockSize: 4096
+        }
+        KeepFileStore: true
+        MaxFileCount: 3
+        RequestsCount: 30
+        IODepth: 16
+        DataLoadSpec {
+            Actions {
+                Action: ACTION_READ
+                Rate: 1
+            }
+            ReadBytes: 4096
+            WriteBytes: 4096
+            MinFileCount: 3
+        }
+    }
+}
