+++ cloud/filestore/tests/loadtest/service-kikimr-test/test.py
@@ -2,6 +2,7 @@ import os
 import pytest
 
 from cloud.filestore.tests.python.lib.loadtest import run_load_test
+from cloud.filestore.tests.python.lib.client import FilestoreCliClient
 
 import yatest.common as common
 
@@ -51,3 +52,35 @@ def test_load(test_case):
     )
 
     return None
+
+
+def test_max_file_count():
+    filesystem_id = "max-file-count"
+    config_path = common.source_path(
+        "cloud/filestore/tests/loadtest/service-kikimr-test/max-file-count.txt"
+    )
+
+    try:
+        run_load_test(
+            filesystem_id,
+            config_path,
+            os.getenv("NFS_SERVER_PORT"),
+        )
+
+        client = FilestoreCliClient(
+            common.binary_path("cloud/filestore/apps/client/filestore-client"),
+            os.getenv("NFS_SERVER_PORT"),
+            cwd=common.output_path(),
+        )
+        entries = client.find(filesystem_id, depth=1).decode().splitlines()
+        entries = [entry for entry in entries if entry]
+
+        assert len(entries) == 3
+    finally:
+        client = FilestoreCliClient(
+            common.binary_path("cloud/filestore/apps/client/filestore-client"),
+            os.getenv("NFS_SERVER_PORT"),
+            cwd=common.output_path(),
+            check_exit_code=False,
+        )
+        client.destroy(filesystem_id)
