+++ cloud/filestore/tests/loadtest/service-kikimr-test/ya.make
@@ -7,6 +7,7 @@ TEST_SRCS(
 )
 
 DEPENDS(
+    cloud/filestore/apps/client
     cloud/filestore/tools/testing/loadtest/bin
 )
